requirements.txt
flask
flask-sqlalchemy
python-dotenv
werkzeug
flask-migrate
